Split up the main gui test file
