+++ tests/reference/guiEditor_Main_Final_nwProject.nwx
