+++ tests/reference/guiEditor_Main_Initial_nwProject.nwx
